server/github/tasks: cleanup utils
Inline or co-locate "utils" to where they are called. A function that's only used once does not need to live in a "utils" package.
